requirements.txt
# ── Core deep learning ───────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Scientific computing ──────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0
scikit-learn>=1.2.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.6.0
Pillow>=9.4.0

# ── Progress bars ─────────────────────────────────────────────────────────────
tqdm>=4.65.0

# ── Time-series baseline ──────────────────────────────────────────────────────
statsmodels>=0.14.0

# ── JSOC / HMI data access (for prepare_dataset.py) ──────────────────────────
drms>=0.6.0
astropy>=5.2.0
